Compiles pat rec shift local

+++ src/EUTelGeometryTelescopeGeoDescription.cpp
@@ -65,7 +65,7 @@ Eigen::Vector3d EUTelGeometryTelescopeGeoDescription::siPlaneNormalEig( int plan
     return norEigen;
 }
 
-std::vector<double> EUTelGeometryTelescopeGeoDescription::localDistDUT() {
+std::vector<float> EUTelGeometryTelescopeGeoDescription::localDistDUT() {
 	return _localDistDUT;
 }
